tests/wpt/web-platform-tests/dom/nodes/ParentNode-querySelector-scope.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>querySelector(All) must work with :scope</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<!-- Regression test for https://github.com/jsdom/jsdom/issues/2359 -->

<div><p><span>hello</span></p></div>

<script>
"use strict";
const div = document.querySelector("div");
const p = document.querySelector("p");

test(() => {
  assert_equals(div.querySelector(":scope > p"), p);
  assert_equals(div.querySelector(":scope > span"), null);
}, "querySelector");

test(() => {
  assert_array_equals(div.querySelectorAll(":scope > p"), [p]);
  assert_array_equals(div.querySelectorAll(":scope > span"), []);
}, "querySelectorAll");
</script>
